+++ src/components/admin/matching/MatchingResultTable.tsx
@@ -1,0 +1,5 @@
+const MatchingResultTable = () => {
+  return <div>MatchingResultTable</div>;
+};
+
+export default MatchingResultTable;
